Include language code in backlink for static and journey pages
