+++ packages/gafl-webapp-service/src/pages/guidance/accessibility-statement.njk
@@ -8,7 +8,7 @@
     <div class="govuk-grid-column-full">
         {{ govukBackLink({
           text: mssgs.back,
-          href: "/buy",
+          href: '/buy' + langCode,
           classes: "govuk-!-margin-bottom-7"
         }) }}
