+++ packages/gafl-webapp-service/src/pages/guidance/cookies.njk
@@ -9,7 +9,7 @@
     <div class="govuk-grid-column-full">
         {{ govukBackLink({
           text: mssgs.back,
-          href: "/buy",
+          href: '/buy' + langCode,
           classes: "govuk-!-margin-bottom-7"
         }) }}
