+++ packages/gafl-webapp-service/src/pages/guidance/os-terms.njk
@@ -8,7 +8,7 @@
     <div class="govuk-grid-column-two-thirds">
         {{ govukBackLink({
           text: mssgs.back,
-          href: "/buy",
+          href: '/buy' + langCode,
           classes: "govuk-!-margin-bottom-7"
         }) }}
