+++ packages/gafl-webapp-service/src/pages/guidance/privacy-policy.njk
@@ -9,7 +9,7 @@
 
         {{ govukBackLink({
           text: mssgs.back,
-          href: "/buy",
+          href: '/buy' + langCode,
           classes: "govuk-!-margin-bottom-7"
         }) }}
