+++ packages/gafl-webapp-service/src/pages/guidance/refund-policy.njk
@@ -8,7 +8,7 @@
     <div class="govuk-grid-column-full">
         {{ govukBackLink({
           text: mssgs.back,
-          href: "/buy",
+          href: '/buy' + langCode,
           classes: "govuk-!-margin-bottom-7"
         }) }}
         <h1 class="govuk-heading-l">{{ mssgs.refund_heading }}</h1>
